Update common declarations syntax
The new syntax came from the fact that Prolog is dynamically typed.

Jing doesn't care or need to know what types the arguments to
actions, fluents or prolog predicates have, it just needs to know the
nubmer of arguments so at a later stage we can enforce simple type
checking on the argument count.

closes #8
/cc #5