Translate `search` statement
Initial handling of `search`. While doing this it
made sense to move the code generation out of
`main` and into its own module `emitter`, which
is an attempt to begin generalising the code
generation.

Semantic checks for `action` calls within a
`search` block is yet to be implemented.

Part of #11