Move semantic checking to its own module
Similar to code emission, the semantic checking has been moved to its
own module.

The semantic checker in its current form halts program execution on the
first error it comes across (there is only one error at present).
Granted this is a simple approach but all that is needed for now.

A side effect of these changes is that semantic checks for `action` are
performed for `search` blocks and possibly other statments too
(implementation and tests needed to verify this).

Part of #11.