Update tpass to check generated code
Passing tests now require a comments section that specifies the expected
output of the generated Prolog code.

The expected output is marked in the test file with a comment:

```
// EXPECTED OUTPUT
// expected code ...
```

The tpass program is very strict about how it finds expected comment
sections in a test file. Expected comments must start with two forward
slashes followed by a space character (`// `).