Check for undefined atom in expressions
A dummy type was introduced for procedure arguments when delcaring a
procedure prototype to prevent procedure arguments from being classed as
an undefined error within an expression.

This semantic check aims to cover the following IndiGolog cases:

Example of an error caused from using an undefined item in an
expression. Note that `active` is undefined and allowed as an argument
to somthing (it's an atom):

```
% assume `up` is an action that is already defined
proc(control, [while(myatom, up), set_state(active)]).
% ERROR: holds/2: Undefined procedure: myatom/0
```

Singular example:

```
proc(control, ?(myatom)).
% ERROR: holds/2: Undefined procedure: myatom/0
```

Operator example:

```
proc(control, ?(myatom > 3)).
% ERROR: >/2: Arithmetic: `myatom/0' is not a function
```