Check operands for undefined and callables
Symbol names are reported as 'unknown (FIXME)'.

The problem with `semcheck_errorl` is the offending item is reported as
'unknown'. The limitation in this case is the expression data structure
needs to be traversed to find the offending item. More work is needed to
provide a custom semcheck error data structure that can collect all of
the relevant information at the time of the error instead of traversing
data structures at the time of displaying the error. Going further the
custom errors could be stored in an array or queue and then dumped out
after parsing is done or, at any time that is convenient.