Translate logic expressions
These changes implement the translation of `~`, `&&` and `||` logic
operators.

Operator types were introduced to make it easier to determine which
operator is in a `node_op` for code generation and semantic checking.
At this stage all Prolog operators are given the same operator type
which is all that is needed for now.

Parentheses are nolonger stripped, they are now retained.

IndiGolog operators `neg`, `and`, `or` have higher precedence over the
Prolog operators. For example `?(~n > 1)` and `?(~(n > 1))` are
semantically equivalent and output as `neg(n > 1)` and `?(neg((n > 1))`
respectively.

Includes renaming of the following:

* `node_expr` -> `node_op`, so the emitter and node modules use the same
  naming conventions
* `operator` -> `op_str`