Change `pconc` syntax to use `then` keyword
The prioritised concurrency construct `pconc(p1, p2)` executes in the
following manner: program `p1` has higher priority than program `p2`
and, `p2` may only execute when `p1` is done or blocked.

For this reason it made sense to syntactically reflect this in the Jing
language by introducing the `then` keyword which gives the reader more
meaning of the intention of the `pconc` construct.