Output `pick` variables with `p_` prefix
A new type for pick variables was added to identify symbol references
that are `pick` variables and should be output with the prefix.

Symbol name conflicts were resolved by `pick` variables now including
the `#` as part of their symbol name so they don't clash with other
symbols (and ultimately overwrite type information in an existing
symbol).

For example the following would cause a symbol clash and result in an
invalid IndiGolog program:

```
action a: 0;

procedure foo() {
    pick #a {
        a(#a);
    }
}
```

producing the invalid program that looses its reference to the action
`a`:

```
proc(foo, [pi(p_a, [p_a(p_a)])]).
```